[Hotfix] Создание портала перенесено в компонент Modal. Лишний компонент Portal удалён.
